Module amethyst_network::simulation

Module containing various utilities to run a client/server-based network simulation. Expect more utilities to make their way into this module. e.g. "Component synchronization", "Matchmaking", etc.

Modules

laminar
Network systems implementation backed by the Laminar network protocol.

tcp
Network systems implementation backed by the TCP network protocol.

udp
Network systems implementation backed by the UDP network protocol.

Structs

Message
Structure used to hold message payloads before they are consumed and sent by an underlying NetworkSystem.

NetworkSimulationTime
Resource to track the state of the network simulation separately from the ECS frame timings

NetworkSimulationTimeSystem
This system is used exclusively to update the state of the NetworkSimulationTime resource.

TransportResource
Resource serving as the owner of the queue of messages to be sent. This resource also serves as the interface for other systems to send messages.

Enums

DeliveryRequirement
Specification of the desired delivery guarantee on a message. All examples will use the following: 1, 2, 3, 4, 5, 6 sent from the server. 5, 1, 4, 2, 3 received by the client. Packet 6 was lost on initial send.

NetworkSimulationEvent
Events which can be received from the network.

UrgencyRequirement
Specification of urgency of the sending of a message. Typically we'll want to send messages on simulation tick but the option to send messages immediately is available.