Module amethyst_network::simulation::laminar

Network systems implementation backed by the Laminar network protocol.

Structs

LaminarConfig
Contains the configuration options to configure laminar for special use-cases.

LaminarNetworkBundle
Use this network bundle to add the laminar transport layer to your game.

LaminarSocket
A reliable UDP socket implementation with configurable reliability and ordering guarantees.

LaminarSocketResource
Resource that owns the Laminar socket.

Enums

ErrorKind
Enum with all possible network errors that could occur.