Struct amethyst_network::NetIdentity

pub struct NetIdentity {
    pub uuid: Uuid,
}

A network identity. It can represent either a client or a server. It represents anything that can own an entity or a component. Think of it as an identity card. When used as a resource, it designates the local network uuid.

Fields
uuid: Uuid

The uuid identifying this NetIdentity.