Struct amethyst_network::NetSocketSystem

pub struct NetSocketSystem<E: 'static> where
    E: PartialEq,  {
    pub filters: Vec<Box<NetFilter<E>>>,
    // some fields omitted
}

The System managing the network state and connections. The T generic parameter corresponds to the network event type. Receives events and filters them. Received events will be inserted into the NetReceiveBuffer resource. To send an event, add it to the NetSendBuffer resource.

If both a connection (Connect or Connected) event is received at the same time as another event from the same connection, only the connection event will be considered and rest will be filtered out.

Fields
filters: Vec<Box<NetFilter<E>>>

The list of filters applied on the events received.

Methods

impl<E> NetSocketSystem<E> where
E: Serialize + PartialEq + Send + 'static,

pub fn new(
addr: SocketAddr,
filters: Vec<Box<NetFilter<E>>>
) -> Result<NetSocketSystem<E>, Error>

Creates a NetSocketSystem and binds the Socket on the ip and port added in parameters.